+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added listening_from_me option in Set Typebot
 * Added variables options in Start Typebot
 * Added webhooks for typebot events
+* Added ChamaAI integration
 
 ### Fixed